Improve shell scripts style
Former-commit-id: 63ad10beee0e35b3578e30c985a4e9fe732a4637

+++ bin/get_seroba_db.sh
@@ -5,13 +5,13 @@
 
 # Assume up-to-date if JSON passes checks and the host cannot be resolved to allow offline usage
 
-if  [ ! -f ${DB_LOCAL}/${JSON_FILE} ] || \
-    [ ! "$(grep 'git' ${DB_LOCAL}/${JSON_FILE} | sed -r 's/.+: "(.*)",?/\1/')" == "${DB_REMOTE}" ] || \
-    [ ! "$(grep 'kmer' ${DB_LOCAL}/${JSON_FILE} | sed -r 's/.+: "(.*)",?/\1/')" == "${KMER}" ] || \
-    !((git -C ${DB_LOCAL} pull || echo 'Already up-to-date') | grep -q 'Already up[- ]to[- ]date'); then
+if  [ ! -f "${DB_LOCAL}"/"${JSON_FILE}" ] || \
+    [ ! "$(grep 'git' "${DB_LOCAL}"/"${JSON_FILE}" | sed -r 's/.+: "(.*)",?/\1/')" == "${DB_REMOTE}" ] || \
+    [ ! "$(grep 'kmer' "${DB_LOCAL}"/"${JSON_FILE}" | sed -r 's/.+: "(.*)",?/\1/')" == "${KMER}" ] || \
+    ! ( (git -C "${DB_LOCAL}" pull || echo 'Already up-to-date') | grep -q 'Already up[- ]to[- ]date' ); then
 
-    rm -rf ${DB_LOCAL}/{,.[!.],..?}*
-    git clone ${DB_REMOTE} ${DB_LOCAL}
+    rm -rf "${DB_LOCAL:?}"/{,.[!.],..?}*
+    git clone "${DB_REMOTE}" "${DB_LOCAL}"
 
     CREATE_DB=true
